Core: Heap enhancements, NanoStrings, Fixed*Map, fix in Socket - non-ASCII char < 0 issue
git-svn-id: svn://ultimatepp.org/upp/trunk@6661 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Inet.h
@@ -143,7 +143,7 @@ class TcpSocket {
 	int                     Get_();
 	int                     Peek_();
 	int                     Peek_(int end_time);
-	int                     Peek(int end_time)          { return ptr < end ? *ptr : Peek_(end_time); }
+	int                     Peek(int end_time)          { return ptr < end ? (byte)*ptr : Peek_(end_time); }
 	bool                    IsGlobalTimeout();
 
 	void                    Reset();
@@ -200,9 +200,9 @@ public:
 	bool            WaitRead()                               { return Wait(WAIT_READ); }
 	bool            WaitWrite()                              { return Wait(WAIT_WRITE); }
 
-	int             Peek()                                   { return ptr < end ? *ptr : Peek_(); }
+	int             Peek()                                   { return ptr < end ? (byte)*ptr : Peek_(); }
 	int             Term()                                   { return Peek(); }
-	int             Get()                                    { return ptr < end ? *ptr++ : Get_(); }
+	int             Get()                                    { return ptr < end ? (byte)*ptr++ : Get_(); }
 	int             Get(void *buffer, int len);
 	String          Get(int len);
